+++ test/test_imgaug.py
@@ -4636,8 +4636,14 @@ def test_Polygon_draw_on_image():
     # simple drawing of square
     poly = ia.Polygon([(2, 2), (8, 2), (8, 8), (2, 8)])
     image_poly = poly.draw_on_image(image,
-                                    color=[32, 128, 32], color_perimeter=[0, 255, 0],
-                                    alpha=1.0, alpha_perimeter=1.0,
+                                    color=[32, 128, 32],
+                                    color_fill=[32, 128, 32],
+                                    color_perimeter=[0, 255, 0],
+                                    color_points=[0, 255, 0],
+                                    alpha=1.0,
+                                    alpha_fill=1.0,
+                                    alpha_perimeter=1.0,
+                                    alpha_points=0.0,
                                     raise_if_out_of_image=False)
     assert image_poly.dtype.type == np.uint8
     assert image_poly.shape == (10, 10, 3)
@@ -4650,11 +4656,38 @@ def test_Polygon_draw_on_image():
     expected = np.tile(np.uint8([32, 128, 32]).reshape((1, 1, 3)), (5, 5, 1))
     assert np.all(image_poly[3:8, 3:8, :] == expected)
 
+    # simple drawing of square, use only "color" arg
+    poly = ia.Polygon([(2, 2), (8, 2), (8, 8), (2, 8)])
+    image_poly = poly.draw_on_image(image,
+                                    color=[0, 255, 0],
+                                    alpha=1.0,
+                                    alpha_fill=1.0,
+                                    alpha_perimeter=1.0,
+                                    alpha_points=0.0,
+                                    raise_if_out_of_image=False)
+    assert image_poly.dtype.type == np.uint8
+    assert image_poly.shape == (10, 10, 3)
+    assert np.sum(image) == 3 * np.sum(np.arange(100))  # draw did not change original image (copy=True)
+    for c_idx, value in enumerate([0, 0.5*255, 0]):
+        value = int(value)
+        assert np.all(image_poly[2:9, 2:3, c_idx] == np.zeros((7, 1), dtype=np.uint8) + value)  # left boundary
+        assert np.all(image_poly[2:9, 8:9, c_idx] == np.zeros((7, 1), dtype=np.uint8) + value)  # right boundary
+        assert np.all(image_poly[2:3, 2:9, c_idx] == np.zeros((1, 7), dtype=np.uint8) + value)  # top boundary
+        assert np.all(image_poly[8:9, 2:9, c_idx] == np.zeros((1, 7), dtype=np.uint8) + value)  # bottom boundary
+    expected = np.tile(np.uint8([0, 255, 0]).reshape((1, 1, 3)), (5, 5, 1))
+    assert np.all(image_poly[3:8, 3:8, :] == expected)
+
     # simple drawing of square with float32 input
     poly = ia.Polygon([(2, 2), (8, 2), (8, 8), (2, 8)])
     image_poly = poly.draw_on_image(image.astype(np.float32),
-                                    color=[32, 128, 32], color_perimeter=[0, 255, 0],
-                                    alpha=1.0, alpha_perimeter=1.0,
+                                    color=[32, 128, 32],
+                                    color_fill=[32, 128, 32],
+                                    color_perimeter=[0, 255, 0],
+                                    color_points=[0, 255, 0],
+                                    alpha=1.0,
+                                    alpha_fill=1.0,
+                                    alpha_perimeter=1.0,
+                                    alpha_points=0.0,
                                     raise_if_out_of_image=False)
     assert image_poly.dtype.type == np.float32
     assert image_poly.shape == (10, 10, 3)
@@ -4669,8 +4702,14 @@ def test_Polygon_draw_on_image():
     # drawing of poly that is half out of image
     poly = ia.Polygon([(2, 2+5), (8, 2+5), (8, 8+5), (2, 8+5)])
     image_poly = poly.draw_on_image(image,
-                                    color=[32, 128, 32], color_perimeter=[0, 255, 0],
-                                    alpha=1.0, alpha_perimeter=1.0,
+                                    color=[32, 128, 32],
+                                    color_fill=[32, 128, 32],
+                                    color_perimeter=[0, 255, 0],
+                                    color_points=[0, 255, 0],
+                                    alpha=1.0,
+                                    alpha_fill=1.0,
+                                    alpha_perimeter=1.0,
+                                    alpha_points=0.0,
                                     raise_if_out_of_image=False)
     assert image_poly.dtype.type == np.uint8
     assert image_poly.shape == (10, 10, 3)
@@ -4687,8 +4726,14 @@ def test_Polygon_draw_on_image():
     got_exception = False
     try:
         _ = poly.draw_on_image(image,
-                               color=[32, 128, 32], color_perimeter=[0, 255, 0],
-                               alpha=1.0, alpha_perimeter=1.0,
+                               color=[32, 128, 32],
+                               color_fill=[32, 128, 32],
+                               color_perimeter=[0, 255, 0],
+                               color_points=[0, 255, 0],
+                               alpha=1.0,
+                               alpha_fill=1.0,
+                               alpha_perimeter=1.0,
+                               alpha_points=0.0,
                                raise_if_out_of_image=True)
     except Exception as exc:
         assert "Cannot draw polygon" in str(exc)
@@ -4698,8 +4743,14 @@ def test_Polygon_draw_on_image():
     # drawing of poly that is fully out of image
     poly = ia.Polygon([(100, 100), (100+10, 100), (100+10, 100+10), (100, 100+10)])
     image_poly = poly.draw_on_image(image,
-                                    color=[32, 128, 32], color_perimeter=[0, 255, 0],
-                                    alpha=1.0, alpha_perimeter=1.0,
+                                    color=[32, 128, 32],
+                                    color_fill=[32, 128, 32],
+                                    color_perimeter=[0, 255, 0],
+                                    color_points=[0, 255, 0],
+                                    alpha=1.0,
+                                    alpha_fill=1.0,
+                                    alpha_perimeter=1.0,
+                                    alpha_points=0.0,
                                     raise_if_out_of_image=False)
     assert np.array_equal(image_poly, image)
 
@@ -4708,8 +4759,14 @@ def test_Polygon_draw_on_image():
     got_exception = False
     try:
         _ = poly.draw_on_image(image,
-                               color=[32, 128, 32], color_perimeter=[0, 255, 0],
-                               alpha=1.0, alpha_perimeter=1.0,
+                               color=[32, 128, 32],
+                               color_fill=[32, 128, 32],
+                               color_perimeter=[0, 255, 0],
+                               color_points=[0, 255, 0],
+                               alpha=1.0,
+                               alpha_fill=1.0,
+                               alpha_perimeter=1.0,
+                               alpha_points=0.0,
                                raise_if_out_of_image=True)
     except Exception as exc:
         assert "Cannot draw polygon" in str(exc)
@@ -4719,8 +4776,14 @@ def test_Polygon_draw_on_image():
     # face invisible via alpha
     poly = ia.Polygon([(2, 2), (8, 2), (8, 8), (2, 8)])
     image_poly = poly.draw_on_image(image,
-                                    color=[32, 128, 32], color_perimeter=[0, 255, 0],
-                                    alpha=0.0, alpha_perimeter=1.0,
+                                    color=[32, 128, 32],
+                                    color_fill=[32, 128, 32],
+                                    color_perimeter=[0, 255, 0],
+                                    color_points=[0, 255, 0],
+                                    alpha=1.0,
+                                    alpha_fill=0.0,
+                                    alpha_perimeter=1.0,
+                                    alpha_points=0.0,
                                     raise_if_out_of_image=False)
     assert image_poly.dtype.type == np.uint8
     assert image_poly.shape == (10, 10, 3)
@@ -4732,8 +4795,14 @@ def test_Polygon_draw_on_image():
     # boundary invisible via alpha
     poly = ia.Polygon([(2, 2), (8, 2), (8, 8), (2, 8)])
     image_poly = poly.draw_on_image(image,
-                                    color=[32, 128, 32], color_perimeter=[0, 255, 0],
-                                    alpha=1.0, alpha_perimeter=0.0,
+                                    color=[32, 128, 32],
+                                    color_fill=[32, 128, 32],
+                                    color_perimeter=[0, 255, 0],
+                                    color_points=[0, 255, 0],
+                                    alpha=1.0,
+                                    alpha_fill=1.0,
+                                    alpha_perimeter=0.0,
+                                    alpha_points=0.0,
                                     raise_if_out_of_image=False)
     assert image_poly.dtype.type == np.uint8
     assert image_poly.shape == (10, 10, 3)
@@ -4741,11 +4810,41 @@ def test_Polygon_draw_on_image():
     expected = np.tile(np.uint8([32, 128, 32]).reshape((1, 1, 3)), (6, 6, 1))
     assert np.all(image_poly[2:8, 2:8, :] == expected)
 
-    # alpha=0.5
+    # alpha=0.8
     poly = ia.Polygon([(2, 2), (8, 2), (8, 8), (2, 8)])
     image_poly = poly.draw_on_image(image,
-                                    color=[32, 128, 32], color_perimeter=[0, 255, 0],
-                                    alpha=0.5, alpha_perimeter=0.5,
+                                    color=[32, 128, 32],
+                                    color_fill=[32, 128, 32],
+                                    color_perimeter=[0, 255, 0],
+                                    color_points=[0, 255, 0],
+                                    alpha=0.8,
+                                    alpha_points=0.0,
+                                    raise_if_out_of_image=False)
+    assert image_poly.dtype.type == np.uint8
+    assert image_poly.shape == (10, 10, 3)
+    for c_idx, value in enumerate([0, 255, 0]):
+        assert np.all(
+            image_poly[2:9, 8:9, c_idx] ==
+            (
+                (1-0.8)*image[2:9, 8:9, c_idx]
+                + np.full((7, 1), 0.8*value, dtype=np.float32)
+            ).astype(np.uint8)
+        )  # right boundary
+    expected = (0.8 * 0.5) * np.tile(np.uint8([32, 128, 32]).reshape((1, 1, 3)), (5, 5, 1)) \
+        + (1 - (0.8 * 0.5)) * image[3:8, 3:8, :]
+    assert np.all(image_poly[3:8, 3:8, :] == expected.astype(np.uint8))
+
+    # alpha of fill and perimeter 0.5
+    poly = ia.Polygon([(2, 2), (8, 2), (8, 8), (2, 8)])
+    image_poly = poly.draw_on_image(image,
+                                    color=[32, 128, 32],
+                                    color_fill=[32, 128, 32],
+                                    color_perimeter=[0, 255, 0],
+                                    color_points=[0, 255, 0],
+                                    alpha=1.0,
+                                    alpha_fill=0.5,
+                                    alpha_perimeter=0.5,
+                                    alpha_points=0.0,
                                     raise_if_out_of_image=False)
     assert image_poly.dtype.type == np.uint8
     assert image_poly.shape == (10, 10, 3)